Add files via upload
Brief Problem Description:
Lightning McQueen and Mator have the used car business in Radiator Springs, but they’re missing one key detail: exact prices for their cars. Instead, they’ve got a broad ‘high’ or ‘low’ price label to work with. They’re counting on us to predict these labels accurately, as this will help them market the cars correctly.
Objective:
The aim is to create a predictive model that can classify used cars into ‘high’ or ‘low’ price categories based on the provided data set. Parts of the data will be used in a training and validation process for two models, and the chosen model will play an essential part in predicting the price categories for a new set of customer data sets, which will have a direct impact on the used car marketing and sales strategy of our client.